requirement.txt
absl-py==2.1.0
affine==2.4.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
anytree==2.12.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
av==14.0.1
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
bokeh==3.6.2
brax==0.10.3
cachetools==5.5.0
Cartopy==0.24.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.88
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contextily==1.6.2
contourpy==1.3.1
cycler==0.12.1
dask==2024.12.1
datashader==0.16.3
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
dm-env==1.6
dm-tree==0.1.8
dotmap==1.3.30
einops==0.8.0
etils==1.11.0
evogym==2.0.0
evosax==0.1.6
exceptiongroup==1.2.2
executing==2.1.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
flax==0.10.2
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.12.0
functional==0.4
future==1.0.0
gast==0.6.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
glfw==2.8.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
gymnax==0.0.8
h11==0.14.0
h5py==3.14.0
holoviews==1.20.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.27.0
humanize==4.11.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.38
jax-cuda12-pjrt==0.4.38
jax-cuda12-plugin==0.4.38
jax-tqdm==0.3.1
jaxlib==0.4.38
jaxmarl==0.0.7
jaxopt==0.8.3
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jupytext==1.16.6
keras==3.11.3
kiwisolver==1.4.8
kmodes==0.12.2
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mercantile==1.2.1
mistune==3.0.2
ml_collections==1.0.0
ml_dtypes==0.5.3
mlx==0.21.1
mlx-lm==0.20.6
moviepy==2.1.1
msgpack==1.1.0
mujoco==3.1.3
mujoco-mjx==3.1.3
multipledispatch==1.0.0
multiprocess==0.70.17
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.6.0.74
nvidia-cufft-cu12==11.3.0.4
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nccl-cu12==2.24.3
nvidia-nvjitlink-cu12==12.6.85
ollama==0.4.5
openai==1.101.0
opencv-contrib-python-headless==4.10.0.84
opencv-python==4.10.0
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
optree==0.17.0
orbax-checkpoint==0.10.3
osmnx==2.0.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
panel==1.5.5
param==2.2.0
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyct==0.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.18.0
pynndescent==0.5.13
pyogrio==0.10.0
PyOpenGL==3.1.7
pyparsing==3.2.0
PyPDF2==3.0.1
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytinyrenderer==0.0.14
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shapely==2.0.6
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.20.0
tensorstore==0.1.71
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.12.12
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trimesh==4.5.3
ttkbootstrap==1.13.8
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
umap-learn==0.5.7
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.3
xarray==2025.1.0
xyzservices==2024.9.0
zipp==3.21.0
